# 改动 - 现在引入了 5.9999968E7D 限制移除。 - 修改了部分表达式与方法名。
# 预期
- 视觉上的蓝色边界墙被移除。

+++ src/main/java/com/inf/farlands/mixin/expand/xz/WorldBorderMixin.java
@@ -29,6 +29,6 @@ public class WorldBorderMixin {
 
     @ModifyConstant(method = "<init>(Lnet/minecraft/world/level/border/WorldBorder$Settings;)V", constant = @Constant(doubleValue = 5.9999968E7D))
     private double OnInitB(double value) {
-        return (double) (FarlandsConstant.MAX_BLOCK - 16) * 2;
+        return (FarlandsConstant.MAX_BLOCK - 16.0) * 2.0;
     }
 }
